fix error migrat

+++ database/migrations/2023_02_24_095604_create_order_details_table.php
@@ -15,7 +15,7 @@ class CreateOrderDetailsTable extends Migration
     {
       Schema::create('order_details', function (Blueprint $table) {
         $table->bigIncrements('id');
-        $table->bigInteger('order_id')->unsigned();
+        $table->integer('order_id')->unsigned();
         $table->foreign('order_id')->references('id')->on('order');
         $table->integer('product_id')->unsigned();
         $table->foreign('product_id')->references('id')->on('products');
